# Build bridge graph to improve performance of bridge shortest path