Most of the time spent in the bridge shortest path algorithm is due to the type instability of bridge types accessed from the vector of bridges added to the `LazyBridgeOptimizer`.
Indeed, we then need to call `supports_constraint`, `concrete_bridge_type` and `added_...` to the bridge type and the method lookup needs to be done at runtime because the bridge type is not known at compile time.
In this PR, the subpart of the bridge graph is constructed so that the Bellman-Ford algorithm is run on a fully type-stable graph structure.
The graph is built lazily when `node` is called for new types.
Of course, the type instability still affects the `node` functions but once the graph is build, the time to run Bellman-Ford is negligible.
Running the `perf/bellman_ford.jl` benchmark gives which shows that solving the shortest path problem takes a negligible time compared to building the graph:
```
BenchmarkTools.Trial: 
  memory estimate:  10.39 KiB
  allocs estimate:  194
  --------------
  minimum time:     237.782 μs (0.00% GC)
  median time:      313.253 μs (0.00% GC)
  mean time:        386.994 μs (2.17% GC)
  maximum time:     54.999 ms (99.18% GC)
  --------------
  samples:          10000
  evals/sample:     1
BenchmarkTools.Trial: 
  memory estimate:  11.80 KiB
  allocs estimate:  284
  --------------
  minimum time:     239.733 μs (0.00% GC)
  median time:      311.878 μs (0.00% GC)
  mean time:        385.405 μs (2.63% GC)
  maximum time:     72.794 ms (99.43% GC)
  --------------
  samples:          10000
  evals/sample:     1
BenchmarkTools.Trial: 
  memory estimate:  25.72 KiB
  allocs estimate:  479
  --------------
  minimum time:     1.614 ms (0.00% GC)
  median time:      1.789 ms (0.00% GC)
  mean time:        2.268 ms (1.41% GC)
  maximum time:     52.869 ms (95.84% GC)
  --------------
  samples:          2200
  evals/sample:     1
BenchmarkTools.Trial: 
  memory estimate:  29.16 KiB
  allocs estimate:  699
  --------------
  minimum time:     1.617 ms (0.00% GC)
  median time:      1.784 ms (0.00% GC)
  mean time:        2.035 ms (1.34% GC)
  maximum time:     50.862 ms (95.61% GC)
  --------------
  samples:          2453
  evals/sample:     1
```
In comparison, the benchmark result was previously (see https://github.com/JuliaOpt/MathOptInterface.jl/pull/945):
```
BenchmarkTools.Trial: 
  memory estimate:  68.59 KiB
  allocs estimate:  1861
  --------------
  minimum time:     3.819 ms (0.00% GC)
  median time:      5.165 ms (0.00% GC)
  mean time:        5.701 ms (2.53% GC)
  maximum time:     110.109 ms (91.74% GC)
  --------------
  samples:          875
  evals/sample:     1
BenchmarkTools.Trial: 
  memory estimate:  163.96 KiB
  allocs estimate:  4440
  --------------
  minimum time:     23.947 ms (0.00% GC)
  median time:      32.318 ms (0.00% GC)
  mean time:        32.113 ms (0.80% GC)
  maximum time:     83.834 ms (47.83% GC)
  --------------
  samples:          156
  evals/sample:     1
```
This gives almost a 20x speedup !