@blegat 2019-11-18
There was a bug in the earlier version of the PR, it is now fixed, I have updated the benchmark, the speedup is now even better.